Move method to compare to BitmapComparer because it was called from several classes
