+++ Tests/Mapsui.Rendering.Skia.Tests/MapRegressionTests.cs
@@ -4,6 +4,7 @@
 
 using Mapsui.Logging;
 using Mapsui.Rendering.Skia.Tests.Helpers;
+using Mapsui.Rendering.Skia.Tests.Utilities;
 using Mapsui.Samples.Common;
 using Mapsui.Samples.Common.Maps.Animations;
 using Mapsui.Samples.Common.Maps.DataFormats;
@@ -129,7 +130,7 @@ public class MapRegressionTests
                     }
                     else
                     {
-                        Assert.That(MapRendererTests.CompareBitmaps(originalStream, bitmap, 1, 0.995), Is.True,
+                        Assert.That(BitmapComparer.Compare(originalStream, bitmap, 1, 0.995), Is.True,
                             $"Fail in sample '{sample.Name}' in category '{sample.Category}'. Image compare failed. The generated image is not equal to the reference image.");
                     }
                 }
