+++ Tests/Mapsui.Rendering.Skia.Tests/MapRendererTests.cs
@@ -1,13 +1,11 @@
-﻿using System;
-using System.IO;
-using System.Threading;
+﻿using System.Threading;
 using System.Threading.Tasks;
 using Mapsui.Rendering.Skia.Tests.Extensions;
+using Mapsui.Rendering.Skia.Tests.Utilities;
 using Mapsui.Samples.Common.Maps.Tests;
 using Mapsui.Styles;
 using Mapsui.Tests.Common.Maps;
 using NUnit.Framework;
-using SkiaSharp;
 
 namespace Mapsui.Rendering.Skia.Tests;
 
@@ -30,7 +28,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
     }
 
     [Test]
@@ -51,7 +49,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
     }
 
     [Test]
@@ -72,7 +70,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
     }
 
     [Test]
@@ -93,7 +91,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
     }
 
     [Test]
@@ -114,7 +112,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
     }
 
     [Test]
@@ -135,7 +133,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
     }
 
     [Test]
@@ -154,7 +152,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
     }
 
     [Test]
@@ -173,7 +171,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
     }
 
     [Test]
@@ -194,7 +192,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap), Is.True);
     }
 
     [Test]
@@ -213,7 +211,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap), Is.True);
     }
 
     [Test]
@@ -232,7 +230,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap), Is.True);
     }
 
     [Test]
@@ -251,7 +249,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
     }
 
     [Test]
@@ -270,7 +268,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
     }
 
     [Test]
@@ -289,7 +287,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
     }
 
     [Test]
@@ -308,7 +306,7 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.995), Is.True);
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.995), Is.True);
     }
 
     [Test]
@@ -327,68 +325,6 @@ internal class MapRendererTests
         File.WriteToGeneratedTestImagesFolder(fileName, bitmap);
 
         // Assert
-        Assert.That(CompareBitmaps(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
-    }
-
-    private static bool CompareColors(SKColor color1, SKColor color2, int allowedColorDistance)
-    {
-        if (color1.Alpha == 0 && color2.Alpha == 0) return true; // If both are transparent all colors are ignored
-        if (Math.Abs(color1.Alpha - color2.Alpha) > allowedColorDistance) return false;
-        if (Math.Abs(color1.Red - color2.Red) > allowedColorDistance) return false;
-        if (Math.Abs(color1.Green - color2.Green) > allowedColorDistance) return false;
-        if (Math.Abs(color1.Blue - color2.Blue) > allowedColorDistance) return false;
-        return true;
-    }
-
-    public static bool CompareBitmaps(Stream? bitmapStream1, Stream? bitmapStream2, int allowedColorDistance = 0, double proportionCorrect = 1)
-    {
-        // The bitmaps in WPF can slightly differ from test to test. No idea why. So introduced proportion correct.
-
-        long trueCount = 0;
-        long falseCount = 0;
-
-        if (bitmapStream1 == null && bitmapStream2 == null)
-        {
-            return true;
-        }
-
-        if (bitmapStream1 == null || bitmapStream2 == null)
-        {
-            return false;
-        }
-
-        bitmapStream1.Position = 0;
-        bitmapStream2.Position = 0;
-
-        using var skData1 = SKData.Create(bitmapStream1);
-        var bitmap1 = SKBitmap.FromImage(SKImage.FromEncodedData(skData1));
-        using var skData2 = SKData.Create(bitmapStream2);
-        var bitmap2 = SKBitmap.FromImage(SKImage.FromEncodedData(skData2));
-
-        if (bitmap1.Width != bitmap2.Width || bitmap1.Height != bitmap2.Height)
-        {
-            return false;
-        }
-
-        for (var x = 0; x < bitmap1.Width; x++)
-        {
-            for (var y = 0; y < bitmap1.Height; y++)
-            {
-                var color1 = bitmap1.GetPixel(x, y);
-                var color2 = bitmap2.GetPixel(x, y);
-                if (color1 == color2)
-                    trueCount++;
-                else
-                {
-                    if (CompareColors(color1, color2, allowedColorDistance))
-                        trueCount++;
-                    else
-                        falseCount++;
-                }
-            }
-        }
-
-        var proportion = (double)(trueCount) / (trueCount + falseCount);
-        return proportionCorrect <= proportion;
+        Assert.That(BitmapComparer.Compare(File.ReadFromOriginalFolder(fileName), bitmap, 1, 0.99), Is.True);
     }
 }
